requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
accelerate>=0.27.0
trl>=0.8.0
peft>=0.10.0
sentencepiece>=0.1.99
Pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
gradio>=4.20.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
tqdm>=4.65.0
datasets>=2.18.0